Updating dependencies
Only one code change was needed to fix a deprecation warning after
upgrading these crates. The motivation for these upgrades are to get
crates that depend on webpki to be updated to depend on rustls-webpki to
address https://rustsec.org/advisories/RUSTSEC-2023-0052. Both
webpki-roots and futures-rustls have released updates that use the new
dependency.